[style] chat -> chatApi로 통합
