+++ src/pages/chatPage/hooks/useQuery/useChatRoomDetail.ts
@@ -1,5 +1,5 @@
 import { useSuspenseQuery } from "@tanstack/react-query";
-import { getChatRoomDetail } from "@/apis/chat";
+import { getChatRoomDetail } from "@/apis/chatApi";
 import type { GetChatRoomDetailResponse } from "@/types/chat";
 
 export const useChatRoomDetail = (chatId: string | undefined) =>
